fix: route github callback through post login
